evaluation/opensource-llms/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argilla==1.10.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.23
click==8.1.3
clickhouse-connect==0.6.3
colorclass==2.2.2
commonmark==0.9.1
compressed-rtf==1.0.6
cryptography==41.0.1
dataclasses-json==0.5.8
Deprecated==1.2.14
duckdb==0.8.1
easygui==0.98.3
ebcdic==1.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
extract-msg==0.41.1
fastapi==0.97.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gpt4all==0.2.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.15.1
idna==3.4
IMAPClient==2.3.1
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.177
lark-parser==0.12.0
llama-cpp-python==0.1.50
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.0.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
oletools==0.60.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pandoc==2.3
pcodedmp==1.2.6
pdfminer.six==20221105
Pillow==9.5.0
plumbum==1.8.2
ply==3.11
posthog==3.0.1
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pypandoc==1.11
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
red-black-tree-mod==1.20
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
RTFDE==0.0.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.2
unstructured==0.6.6
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.14.1
XlsxWriter==3.1.2
yarl==1.9.2
zstandard==0.21.0
# add vicuna
llama-cpp-python==0.1.48
protobuf==3.20.0; sys_platform != 'darwin'
protobuf==3.20.0; sys_platform == 'darwin' and platform_machine != 'arm64'
protobuf==3.20.3; sys_platform == 'darwin' and platform_machine == 'arm64'
auto-gptq


